requirements.txt
django
djangorestframework
mysqlclient
Pillow